@@ -1,17 +1,7 @@
-﻿using System;
-using System.Diagnostics;
-using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.Versioning;
-using System.Security;
-using System.Security.Permissions;
+﻿using System.Reflection;
 
 [assembly: AssemblyVersion("1.0.0.0")]
 [assembly: AssemblyCompany("Codexus.Interceptors")]
-[assembly: AssemblyConfiguration("Debug")]
 [assembly: AssemblyFileVersion("1.0.0.0")]
-[assembly: AssemblyInformationalVersion("1.0.0.0+f48f00e06c29fb5eba2136a02b77278c5aaa27d7")]
 [assembly: AssemblyProduct("Codexus.Interceptors")]
 [assembly: AssemblyTitle("Codexus.Interceptors")]
-[assembly: SecurityPermission(SecurityAction.RequestMinimum, SkipVerification = true)]
-[module: RefSafetyRules(11)]
